This script kills idle, orphaned sshd daemon processes
Bliss leaves sshd processes lingering around, which at this point is hard to
fix.  This script works around that problem by identifying the lingering
processes and killing them.  It can be invoked manually, but can of course also
be run via cron.  To add it to cron, do the following on the target machine:

  - export EDITOR=vim
  - crontab -e
  - at the end of that file, add the line:
    0 * * * * $HOME/kill_sshd.pl
    exit the editor

The above would run the script every full hour.  If you move the '0' to the
second position, the script is run once per day.

I *know* this should be fixed in Bliss, and I *know* this is a hack --
don't rub it in...

A.